mtn_momo_hackathon/MOMO_2.0/backend/requirements.txt
python==3.10.12
uvicorn
python-dotenv
tiktoken
langchain==0.0.329
openai==0.28.1
faiss-cpu
fastapi
openai

